better fix for https://github.com/PacificBiosciences/FALCON/issues/21
change the default behavior for more robust import when the end and
begin coordinate in the fasta file does not match the read length.

If the end - beg matches read lenth, the end and beg will be used in the
DB, otherwise, beg will be set to 0 and end will be set to the read
length.